requirements.txt
Django>=2.0,<3.0
Faker
djangorestframework==3.11.0